Feat : disable call sftp

+++ backend/src/main/java/com/viggo/accor/service/SftpUploadService.java
@@ -78,9 +78,9 @@ public class SftpUploadService {
 
 		uploadGateway.upload(csvOutputFile);
 
-	/*	if (csvOutputFile.exists()) {
+		if (csvOutputFile.exists()) {
 			csvOutputFile.delete();
-		}*/
+		}
 	}
 
 	public void uploadFileToSftp(List<CsvFormatDTO> csvFormats) throws FileNotFoundException {
